Support drop table.

+++ src/storage/table.rs
@@ -62,9 +62,14 @@ impl Table {
         let data_page = self.storage.read(page_idx)?;
         let mut row = vec![];
         let mut curr_offset = offset as usize;
-        let sizes = self.schema.get_sizes();
+        // let sizes = self.schema.get_sizes();
         
-        for size in sizes {
+        for col in self.schema.get_columns() {
+            let size = DataItem::cal_size_from_coltype(&col.data_type);
+            if col.is_dropped {
+                curr_offset += size;
+                continue;
+            }
             let item_bytes = &data_page.data[curr_offset..curr_offset+size];
             match DataItem::from_bytes(item_bytes, None)? {
                 DataItem::VarChar { head, value } => {
@@ -410,41 +415,62 @@ impl Table {
         self.schema.satisfy(&data)?;
         // check unique constraints separately
         // the pk must be unique, so no need to check again
-        for (i, col) in self.schema.get_columns().iter().enumerate() {
+        let mut visible_col_idx = 0;
+        for col in self.schema.get_columns().iter() {
+            if col.is_dropped { continue; }
             if col.unique {
                 let index = self.indexes.get(&col.name).unwrap();
-                let existing = index.check_exists(data[i].clone(), &self.storage)?;
+                let existing = index.check_exists(data[visible_col_idx].clone(), &self.storage)?;
                 if existing {
                     return Err(RsqlError::InvalidInput(
                         format!("Unique constraint violation on column {}", col.name)));
                 }
             }
+            visible_col_idx += 1;
         }
         // 2. allocate entry
         let (entry_page_idx, entry_offset) = self.allocator.alloc_entry(tnx_id, &mut self.storage)?;
-        // 3. store VarChar data if any
-        let mut converted_data = data.clone();
-        for (i, dataitem) in data.into_iter().enumerate() {
-            if let DataItem::VarChar { .. } = dataitem {
-                let col = &self.schema.get_columns()[i];
-                let max_len = match &col.data_type {
-                    crate::catalog::table_schema::ColType::VarChar(max_len) => *max_len,
-                    _ => panic!("Column type mismatch for VarChar"),
-                } as u64;
-                let stored_varchar = self.store_varchar(dataitem, tnx_id, max_len)?;
-                converted_data[i] = stored_varchar;
+        // 3. construct physical data (handling dropped columns and VarChar)
+        let mut physical_data = vec![];
+        let mut input_iter = data.into_iter();
+
+        let columns = self.schema.get_columns().clone();
+        for col in columns {
+            if col.is_dropped {
+                let dummy = match col.data_type {
+                     crate::catalog::table_schema::ColType::Integer => DataItem::NullInt,
+                     crate::catalog::table_schema::ColType::Float => DataItem::NullFloat,
+                     crate::catalog::table_schema::ColType::Bool => DataItem::NullBool,
+                     crate::catalog::table_schema::ColType::Chars(len) => DataItem::NullChars{len: len as u64},
+                     crate::catalog::table_schema::ColType::VarChar(_) => DataItem::NullVarChar,
+                };
+                physical_data.push(dummy);
+            } else {
+                let item = input_iter.next().ok_or(RsqlError::ExecutionError("Not enough data items provided".to_string()))?;
+                
+                // Handle VarChar
+                if let DataItem::VarChar { .. } = &item {
+                    let max_len = match &col.data_type {
+                        crate::catalog::table_schema::ColType::VarChar(max_len) => *max_len,
+                        _ => panic!("Column type mismatch for VarChar"),
+                    } as u64;
+                    let stored_varchar = self.store_varchar(item, tnx_id, max_len)?;
+                    physical_data.push(stored_varchar);
+                } else {
+                    physical_data.push(item);
+                }
             }
         }
         // 4. write entry data
-        let entry_bytes = Self::row_to_bytes(&converted_data)?;
+        let entry_bytes = Self::row_to_bytes(&physical_data)?;
         self.storage.write_bytes(tnx_id, entry_page_idx, entry_offset as usize, &entry_bytes)?;
         // 5. write index entries
         for (i, col) in self.schema.get_columns().iter().enumerate() {
-            if col.index {
+            if col.index && !col.is_dropped {
                 let index = self.indexes.get_mut(&col.name).unwrap();
                 index.insert_entry(
                     tnx_id,
-                    converted_data[i].clone(),
+                    physical_data[i].clone(),
                     entry_page_idx,
                     entry_offset,
                     &mut self.storage,
@@ -478,17 +504,22 @@ impl Table {
             }
         }
         // 5. delete from indexes
-        for (i, col) in self.schema.get_columns().iter().enumerate() {
+        let mut visible_idx = 0;
+        for col in self.schema.get_columns().iter() {
+            if col.is_dropped {
+                continue;
+            }
             if col.index {
                 let index = self.indexes.get_mut(&col.name).unwrap();
                 index.delete_entry(
                     tnx_id,
-                    row[i].clone(),
+                    row[visible_idx].clone(),
                     match_page,
                     match_offset,
                     &mut self.storage,
                 )?;
             }
+            visible_idx += 1;
         }
         Ok(())
     }
@@ -612,6 +643,7 @@ mod tests {
                     nullable: false,
                     index: true,
                     unique: true,
+                    is_dropped: false,
                 },
                 TableColumn {
                     name: "name".to_string(),
@@ -620,6 +652,7 @@ mod tests {
                     nullable: false,
                     index: false,
                     unique: false,
+                    is_dropped: false,
                 },
             ];
         TableSchema::new(columns).unwrap()
@@ -671,6 +704,7 @@ mod tests {
                     nullable: false,
                     index: true,
                     unique: true,
+                    is_dropped: false,
                 },
                 TableColumn {
                     name: "bio".to_string(),
@@ -679,6 +713,7 @@ mod tests {
                     nullable: false,
                     index: false,
                     unique: false,
+                    is_dropped: false,
                 },
             ];
         let schema = TableSchema::new(columns).unwrap();
@@ -826,6 +861,7 @@ mod tests {
                 nullable: false,
                 unique: true,
                 index: true,
+                is_dropped: false,
             },
             crate::catalog::table_schema::TableColumn {
                 name: "age".to_string(),
@@ -834,6 +870,7 @@ mod tests {
                 nullable: false,
                 unique: false,
                 index: false,
+                is_dropped: false,
             },
         ];
         let schema = crate::catalog::table_schema::TableSchema::new(columns).unwrap();
